fix floating point fields in model, add field create feature and basic field details view

+++ AgricultureManagementSystem/Models/Combine.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace AgricultureManagementSystem.Models
 {
@@ -11,7 +9,7 @@ namespace AgricultureManagementSystem.Models
         [Required(ErrorMessage = "Podaj ilość wytrząsaczy")]
         [Display(Name = "Liczba wytrząsaczy")]
         [Range(1, 20, ErrorMessage = "Liczba wytrząsaczy musi mieścić się w przedziale od 1 do 20")]
-        public ushort NumberOfStrawWalkers { get; set; }
+        public byte NumberOfStrawWalkers { get; set; }
 
         public IList<Header> Headers { get; } = new List<Header>();
