+++ AgricultureManagementSystem/Models/Header.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace AgricultureManagementSystem.Models
 {
@@ -15,7 +12,7 @@ namespace AgricultureManagementSystem.Models
         [Required(ErrorMessage = "Podaj szerokość roboczą przystawki")]
         [Display(Name = "Szerokość robocza przystawki (m)")]
         [Range(0.1, 30, ErrorMessage = "Szerokość przystawki musi mieścić się w zakresie od 0.1 do 30")]
-        public ushort Width { get; set; }
+        public float Width { get; set; }
 
         public void Update(Header header)
         {
