+++ AgricultureManagementSystem/Models/Vehicle.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace AgricultureManagementSystem.Models
 {
@@ -11,7 +8,7 @@ namespace AgricultureManagementSystem.Models
         [Required(ErrorMessage = "Podaj przebieg lub motogodziny")]
         [Display(Name = "Przebieg (km / mth)")]
         [Range(1, 10000000, ErrorMessage = "Przebieg musi mieścić się w przedziale od 1 do 10 000 000")]
-        public ushort Course { get; set; }
+        public uint Course { get; set; }
 
         protected Vehicle() { }
     }
